hotel_predict/requirements.txt
jupyter==1.0.0
lxml==4.9.0
matplotlib==3.5.2
pandas==1.4.2
Pillow==9.1.1
scikit-learn==1.1.1
numpy==1.22.4
seaborn==0.11.2
tensorflow==2.9.1
tensorflow-datasets==4.6.0


